+++ Sources/VComponents/Components/Modals/Alert/VAlertUIModel.swift
@@ -47,6 +47,9 @@ public struct VAlertUIModel {
         /// Rounded corners. Set to to `allCorners`.
         public var roundedCorners: RectCorner = .allCorners
         
+        /// Indicates if left and right corners should switch to support RTL languages. Set to `true`.
+        public var reversesLeftAndRightCornersForRTLLanguages: Bool = true
+        
         /// Corner radius. Set to `20`.
         public var cornerRadius: CGFloat = 20
         
@@ -257,6 +260,7 @@ public struct VAlertUIModel {
         var uiModel: VSheetUIModel = .init()
         
         uiModel.layout.roundedCorners = layout.roundedCorners
+        uiModel.layout.reversesLeftAndRightCornersForRTLLanguages = layout.reversesLeftAndRightCornersForRTLLanguages
         uiModel.layout.cornerRadius = layout.cornerRadius
         uiModel.layout.contentMargin = 0