+++ Sources/VComponents/Components/Value Pickers/Slider/VSliderUIModel.swift
@@ -41,6 +41,8 @@ public struct VSliderUIModel {
         public var cornerRadius: CGFloat = GlobalUIModel.Common.barCornerRadius
         
         /// Indicates if slider rounds progress view right-edge. Set to `true`.
+        ///
+        /// For RTL languages, this refers to left-edge.
         public var roundsProgressViewRightEdge: Bool = true
         
         var progressViewRoundedCorners: RectCorner {